nyloswapper: refactor tob-cheats down to just nylo swapper.
