+++ NyloSwapper/pom.xml
@@ -2,8 +2,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <artifactId>Theatre-of-Blood</artifactId>
-    <version>2.6</version>
+	<artifactId>NyloSwapper</artifactId>
+	<version>1.0</version>
     <packaging>jar</packaging>
 
     <parent>
