+++ pom.xml
@@ -17,7 +17,7 @@
 		<module>PraySwapper</module>
 		<module>Scaler</module>
 		<module>SpellCaster</module>
-		<module>Theatre-of-Blood</module>
+		<module>NyloSwapper</module>
 		<module>Tickeater</module>
 		<module>LeftClickPK</module>
 		<module>OneClick</module>
